xfree86: move over public SDK headers to include/
