+++ hw/xfree86/common/xf86Init.c
@@ -52,6 +52,7 @@
 #include "config/hotplug_priv.h"
 #include "dix/input_priv.h"
 #include "dix/screenint_priv.h"
+#include "include/xorgVersion.h"
 #include "mi/mi_priv.h"
 #include "os/cmdline.h"
 #include "os/ddx_priv.h"
@@ -75,7 +76,6 @@
 #include "xf86_os_support.h"
 #include "xf86_OSlib.h"
 #include "xf86cmap.h"
-#include "xorgVersion.h"
 #include "mipointer.h"
 #include "xf86Extensions.h"
 #include "xf86DDC.h"
