+++ hw/xwin/winprocarg.c
@@ -36,7 +36,7 @@ from The Open Group.
 #include "os/osdep.h"
 #include "os/ddx_priv.h"
 
-#include <../xfree86/common/xorgVersion.h>
+#include "include/xorgVersion.h"
 #include "win.h"
 #include "winconfig.h"
 #include "winmsg.h"
