+++ include/xf86Optionstr.h
