+++ include/xf86Priv.h
