+++ include/xf86Privstr.h
