+++ include/xf86cmap.h
