+++ include/xf86fbman.h
