+++ include/xf86str.h
