+++ include/xf86xvmc.h
